ai-engine/requirements.txt
fastapi==0.115.6 
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
google-genai

